Add solution recursively
The recursive approach requires 2 function. the first function takes in an Array and an empty Dictionary. It check to see if the first value is in the dictionary, and increments its value by 1 if not it assigns it the value of 1 before removing it from the array. then it calls its self again, passing in the new array with the updated dictionary. 

The second array just loops through the Dictionary and returns the key that has a value of 1