Create removeDuplicates.swift
This First approach grabs all the values from the LL recursively and inserts them into a Set. Then we turn the set into an Array. looping through the Array, we remove the item from index 0 and creates a Node and asign that node to the previous Node while assigning the head to the newly created Node